// src/generated/live/ObservedMatch.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GameSummary } from "./GameSummary";

export type ObservedMatch = { id: string, observationId: string, platform: string, gameId: string, accountId: string, segmentId: string, firstSeen: number, lastSeen: number, championId: number, queueName: string, state: string, game: GameSummary | null, dataError: string | null, automatic: boolean, manual: boolean, };